+++ backends/bevy_picking_rapier/src/lib.rs
@@ -2,12 +2,13 @@
 //!
 //! # Usage
 //!
-//! This backend requires you mark any cameras that will be used for raycasting with
-//! [`RapierPickCamera`]. If a pointer passes through this camera's render target, it will
+//! If a pointer passes through this camera's render target, it will
 //! automatically shoot rays into the rapier scene and will be able to pick things.
 //!
 //! To ignore an entity, you can add [`Pickable::IGNORE`] to it, and it will be ignored during
 //! raycasting.
+//!
+//! For fine-grained control, see the [`RapierBackendSettings::require_markers`] setting.
 //!
 //! ## Limitations
 //!
@@ -26,11 +27,9 @@
 
 use bevy_app::prelude::*;
 use bevy_ecs::prelude::*;
-use bevy_math::Ray;
-use bevy_reflect::prelude::*;
-use bevy_render::prelude::*;
+use bevy_reflect::{std_traits::ReflectDefault, Reflect};
+use bevy_render::{prelude::*, view::RenderLayers};
 use bevy_transform::prelude::*;
-use bevy_utils::HashMap;
 use bevy_window::PrimaryWindow;
 
 use bevy_picking_core::backend::prelude::*;
@@ -38,7 +37,7 @@ use bevy_rapier3d::prelude::*;
 
 /// Commonly used imports.
 pub mod prelude {
-    pub use crate::{RapierBackend, RapierPickCamera};
+    pub use crate::{RapierBackend, RapierBackendSettings, RapierPickable};
 }
 
 /// Adds the `rapier` raycasting picking backend to your app.
@@ -46,106 +45,105 @@ pub mod prelude {
 pub struct RapierBackend;
 impl Plugin for RapierBackend {
     fn build(&self, app: &mut App) {
-        app.add_systems(
-            PreUpdate,
-            (build_rays_from_pointers, update_hits)
-                .chain()
-                .in_set(PickSet::Backend),
-        );
+        app.init_resource::<RapierBackendSettings>()
+            .add_systems(PreUpdate, update_hits.in_set(PickSet::Backend));
     }
 }
 
-/// Marks a camera that should be used for rapier raycast picking.
+/// Runtime settings for the [`RapierBackend`].
+#[derive(Resource, Default, Reflect)]
+#[reflect(Resource, Default)]
+pub struct RapierBackendSettings {
+    /// When set to `true` raycasting will only happen between cameras and entities marked with
+    /// [`RapierPickable`]. Off by default. This setting is provided to give you fine-grained
+    /// control over which cameras and entities should be used by the rapier backend at runtime.
+    pub require_markers: bool,
+}
+
+/// Optional. Marks cameras and target entities that should be used in the rapier picking backend.
+/// Only needed if [`RapierBackendSettings::require_markers`] is set to true.
 #[derive(Debug, Clone, Default, Component, Reflect)]
 #[reflect(Component, Default)]
-pub struct RapierPickCamera {
-    #[reflect(ignore)]
-    /// Maps the pointers visible to this [`RapierPickCamera`] to their corresponding ray. We need
-    /// to create a map because many pointers may be visible to this camera.
-    ray_map: HashMap<PointerId, Ray>,
-}
+pub struct RapierPickable;
 
-impl RapierPickCamera {
-    /// Returns a map that defines the [`Ray`] associated with every [`PointerId`] that is on this
-    /// [`RapierPickCamera`]'s render target.
-    pub fn ray_map(&self) -> &HashMap<PointerId, Ray> {
-        &self.ray_map
-    }
-}
-
-/// Updates all picking [`Ray`]s with [`PointerLocation`]s.
-pub fn build_rays_from_pointers(
+/// Raycasts into the scene using [`RapierBackendSettings`] and [`PointerLocation`]s, then outputs
+/// [`PointerHits`].
+pub fn update_hits(
     pointers: Query<(&PointerId, &PointerLocation)>,
-    primary_window: Query<Entity, With<PrimaryWindow>>,
-    mut picking_cameras: Query<(&Camera, &GlobalTransform, &mut RapierPickCamera)>,
+    primary_window_entity: Query<Entity, With<PrimaryWindow>>,
+    picking_cameras: Query<(
+        Entity,
+        &Camera,
+        &GlobalTransform,
+        Option<&RapierPickable>,
+        Option<&RenderLayers>,
+    )>,
+    pickables: Query<&Pickable>,
+    marked_targets: Query<&RapierPickable>,
+    layers: Query<&RenderLayers>,
+    backend_settings: Res<RapierBackendSettings>,
+    rapier_context: Option<Res<RapierContext>>,
+    mut output_events: EventWriter<PointerHits>,
 ) {
-    picking_cameras.iter_mut().for_each(|(_, _, mut pick_cam)| {
-        pick_cam.ray_map.clear();
-    });
+    let Some(rapier_context) = rapier_context else {
+        return;
+    };
+
     for (pointer_id, pointer_location) in &pointers {
         let pointer_location = match pointer_location.location() {
             Some(l) => l,
             None => continue,
         };
-        picking_cameras
-            .iter_mut()
-            .filter(|(camera, _, _)| {
-                camera.is_active && pointer_location.is_in_viewport(camera, &primary_window)
+        for (cam_entity, camera, ray, cam_layers) in picking_cameras
+            .iter()
+            .filter(|(_, camera, ..)| {
+                camera.is_active && pointer_location.is_in_viewport(camera, &primary_window_entity)
             })
-            .for_each(|(camera, transform, mut source)| {
+            .filter(|(.., marker, _)| marker.is_some() || !backend_settings.require_markers)
+            .filter_map(|(entity, camera, transform, _, layers)| {
                 let mut viewport_pos = pointer_location.position;
                 if let Some(viewport) = &camera.viewport {
                     viewport_pos -= viewport.physical_position.as_vec2();
                 }
-                if let Some(ray) = camera.viewport_to_world(transform, viewport_pos) {
-                    source.ray_map.insert(*pointer_id, ray);
-                }
-            });
-    }
-}
-
-/// Produces [`PointerHits`]s from [`RapierPickRay`] intersections.
-fn update_hits(
-    rapier_context: Option<Res<RapierContext>>,
-    targets: Query<&Pickable>,
-    mut sources: Query<(Entity, &Camera, &mut RapierPickCamera)>,
-    mut output: EventWriter<PointerHits>,
-) {
-    let rapier_context = match rapier_context {
-        Some(c) => c,
-        None => return,
-    };
-
-    sources
-        .iter_mut()
-        .flat_map(|(entity, camera, source)| {
-            source
-                .ray_map
-                .iter()
-                .map(|(pointer, ray)| (entity, camera.order, *pointer, *ray))
-                .collect::<Vec<_>>()
-        })
-        .filter_map(|(cam_entity, cam_order, pointer, ray)| {
-            rapier_context
+                camera
+                    .viewport_to_world(transform, viewport_pos)
+                    .map(|ray| (entity, camera, ray, layers))
+            })
+        {
+            if let Some((entity, hit_data)) = rapier_context
                 .cast_ray_and_get_normal(
                     ray.origin,
                     ray.direction,
                     f32::MAX,
                     true,
                     QueryFilter::new().predicate(&|entity| {
-                        targets
+                        let marker_requirement =
+                            !backend_settings.require_markers || marked_targets.get(entity).is_ok();
+                        let render_layers_match = match (cam_layers, layers.get(entity)) {
+                            (Some(cam_layers), Ok(entity_layers)) => {
+                                cam_layers.intersects(entity_layers)
+                            }
+                            _ => true, // If either `RenderLayers` components is not present, ignore.
+                        };
+                        let pickable = pickables
                             .get(entity)
-                            .map(|pickable| *pickable != Pickable::IGNORE)
-                            .unwrap_or(true)
+                            .map(|p| *p != Pickable::IGNORE)
+                            .unwrap_or(true);
+                        marker_requirement && render_layers_match && pickable
                     }),
                 )
-                .map(|(target, intersection)| {
-                    (cam_entity, cam_order, pointer, target, intersection)
+                .map(|(entity, hit)| {
+                    let hit_data =
+                        HitData::new(cam_entity, hit.toi, Some(hit.point), Some(hit.normal));
+                    (entity, hit_data)
                 })
-        })
-        .for_each(|(cam_entity, cam_order, pointer, target, hit)| {
-            let hit = HitData::new(cam_entity, hit.toi, Some(hit.point), Some(hit.normal));
-            let order = cam_order as f32;
-            output.send(PointerHits::new(pointer, vec![(target, hit)], order));
-        });
+            {
+                output_events.send(PointerHits::new(
+                    *pointer_id,
+                    vec![(entity, hit_data)],
+                    camera.order as f32,
+                ));
+            }
+        }
+    }
 }